requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
werkzeug==3.0.1